Create a function game() to play 5 rounds
Before creating the function game(), we create 3 variables outside the
function in order to keeps score. The function game() with the parameter
rounds is a recursive one (given that we don't use loops in that stage).
It allows us to play 5 rounds if we specify game(5). For the computer
choice we create a const which uses the function getComputerChoice() and
a const which allows us to get input from th user.